requirements/devl.txt
bpython==0.14.2
django-debug-toolbar==1.4
